# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Note-microservice",
        "version": "1.0.0"
    },
    "paths": {
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a list of notes",
                "operationId": "9a95bde5931986ad60b9f7f414ec425e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Note"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create a new note",
                "operationId": "e28ffd9dcbbcb4e1b04a525ab9c13a01",
                "requestBody": {
                    "description": "Note data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Note"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Get a specific note",
                "operationId": "4f26c25d4040d0ef0fa8d8a5ea9fa4da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update a specific note",
                "operationId": "c7e6224fadd6d82b81ded7615279a15c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated note data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Note"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Note"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete a specific note",
                "operationId": "99b63679b2822651f8f0ae585f927f0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Note deleted successfully"
                    },
                    "404": {
                        "description": "Note not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Note": {
                "required": [
                    "title",
                    "content"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Note Title"
                    },
                    "content": {
                        "type": "string",
                        "example": "Note Content"
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    }
}